Refactor to make routes in app.ts easier to grasp
