+++ packages/github-lens-api/src/app.ts
@@ -3,7 +3,11 @@ import cors from 'cors';
 import type { Express } from 'express';
 import express, { Router } from 'express';
 import asyncHandler from 'express-async-handler';
-import { getDescribeRouterHandler } from '../../common/src/expressRoutes';
+import {getRouteDescriptions} from "./controller/DescribeApiController";
+import {getHealthCheckHandler} from "./controller/HealthCheckController";
+import {getAllMembers, getRepoByLogin} from "./controller/MembersController";
+import {getAllRepos, getRepoByName} from "./controller/ReposController";
+import {getTeamBySlug} from "./controller/TeamsController";
 import type { GitHubData } from './data'
 
 export function buildApp(
@@ -21,9 +25,7 @@ export function buildApp(
 		}),
 	);
 
-	router.get('/healthcheck', (req: express.Request, res: express.Response) => {
-		res.status(200).json({ status: 'OK', stage: 'INFRA' });
-	});
+	router.get('/healthcheck', getHealthCheckHandler(router));
 
 	router.get(
 		'/repos',
@@ -33,21 +35,7 @@ export function buildApp(
 				res.status(500).json({ error: 'Unable to retrieve repository data!'});
 				return;
 			}
-
-			if (typeof req.query.name !== 'undefined') {
-				const searchString:string = req.query.name.toString()
-				const jsonResponse = reposData.payload.filter((item) =>
-					item.name.match(searchString),
-				);
-				if (jsonResponse.length !== 0) {
-					res.status(200).json(jsonResponse);
-				} else {
-					res
-						.status(200)
-						.json({ searchString: searchString, info: 'no results found in repos' });
-				}			} else {
-				res.status(200).json(reposData);
-			}
+				getAllRepos(req, res, reposData);
 		}),
 	);
 
@@ -59,17 +47,7 @@ export function buildApp(
 				res.status(500).json({ error: 'Unable to retrieve repository data!'});
 				return;
 			}
-
-			const jsonResponse = reposData.payload.filter(
-				(item) => item.name === req.params.name,
-			);
-			if (jsonResponse.length !== 0) {
-				res.status(200).json(jsonResponse);
-			} else {
-				res
-					.status(200)
-					.json({ repoName: req.params.name, info: 'Repository not found' });
-			}
+				getRepoByName(req, res, reposData);
 		}),
 	);
 
@@ -81,7 +59,6 @@ export function buildApp(
 				res.status(500).json({ error: 'Unable to retrieve teams data!'});
 				return;
 			}
-
 			res.status(200).json(teamsData);
 		}),
 	);
@@ -94,17 +71,7 @@ export function buildApp(
 				res.status(500).json({ error: 'Unable to retrieve teams data!'});
 				return;
 			}
-
-			const jsonResponse = teamsData.payload.filter(
-				(item) => item.slug === req.params.slug,
-			);
-			if (jsonResponse.length !== 0) {
-				res.status(200).json(jsonResponse);
-			} else {
-				res
-					.status(200)
-					.json({ teamSlug: req.params.slug, info: 'Team not found' });
-			}
+			getTeamBySlug(req, res, teamsData)
 		}),
 	);
 
@@ -116,8 +83,7 @@ export function buildApp(
 				res.status(500).json({ error: 'Unable to retrieve members data!'});
 				return;
 			}
-
-			res.status(200).json(membersData);
+			getAllMembers(req, res, membersData)
 		}),
 	);
 
@@ -129,54 +95,12 @@ export function buildApp(
 				res.status(500).json({ error: 'Unable to retrieve members data!'});
 				return;
 			}
-
-			const jsonResponse = membersData.payload.filter(
-				(item) => item.login === req.params.login,
-			);
-			if (jsonResponse.length !== 0) {
-				res.status(200).json(jsonResponse);
-			} else {
-				res
-					.status(200)
-					.json({ memberLogin: req.params.login, info: 'Member not found' });
-			}
+			getRepoByLogin(req, res, membersData)
 		}),
 	);
 
 	//handle all invalid routes by showing all available routes
-	router.get('*', getDescribeRouterHandler(router, (path: string) => {
-		let info = '';
-		switch (path) {
-			case '/healthcheck':
-				info = 'Display healthcheck';
-				break;
-			case '/repos':
-				info =
-					'Show all repos, with their team owners, optionally search by name with ?name=^repo-name-regex.*';
-				break;
-			case '/repos/:name':
-				info = 'Show repo and its team owners, if it exists';
-				break;
-			case '/teams':
-				info =
-					'Show all teams, with the repositories they own';
-				break;
-			case '/teams/:slug':
-				info = 'Show team and the repositories it owns, if it exists';
-				break;
-			case '/members':
-				info =
-					'Show member, with the teams they are in';
-				break;
-			case '/members/:login':
-				info = 'Show member and the teams they are in, if it exists';
-				break;
-			default:
-				info = 'No path info supplied';
-				break;
-		}
-		return info;
-	}));
+	router.get('*', getRouteDescriptions(router));
 
 	app.use('/', router);
