+++ packages/github-lens-api/src/controller/HealthCheckController.ts
@@ -1,0 +1,8 @@
+import type {Router} from "express";
+import type express from "express";
+
+export const getHealthCheckHandler = (router: Router) => {
+    return (req: express.Request, res: express.Response) => {
+        res.status(200).json({ status: 'OK', stage: 'INFRA' });
+    };
+};
